requirements.txt
asgiref==3.6.0
captcha==0.4
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
django-embed-video==1.4.8
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
idna==3.4
Pillow==9.4.0
PyMySQL==1.0.2
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
